feat(domain): 엔티티 아이디 생성을 추상객체에게 위임해 아이디 생성 전략을 일관성있게 가져감
